:package: Refactor `pythonforandroid.python` recipes into his own recipe
As it was done originally because with only one python recipe, there is
no need to have a subclass.
